Refactor | #163 | @lcomment | listener 패키지 리팩토링
