+++ cakk-api/src/main/java/com/cakk/api/listener/ViewsIncreaseEventListener.java
@@ -1,4 +1,4 @@
-package com.cakk.api.listener.views;
+package com.cakk.api.listener;
 
 import java.util.Objects;
