+++ cakk-api/src/test/java/com/cakk/api/listener/EmailSendEventListenerTest.java
@@ -11,7 +11,6 @@ import net.jqwik.api.Arbitraries;
 import com.cakk.api.common.annotation.TestWithDisplayName;
 import com.cakk.api.common.base.MockitoTest;
 import com.cakk.api.dto.event.EmailWithVerificationCodeSendEvent;
-import com.cakk.api.listener.user.EmailSendEventListener;
 import com.cakk.external.service.MailService;
 
 class EmailSendEventListenerTest extends MockitoTest {
